+++ src/main/java/com/github/wintersteve25/tau/Tau.java
@@ -1,5 +1,6 @@
 package com.github.wintersteve25.tau;
 
+import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.common.Mod;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -9,6 +10,6 @@ public class Tau {
     public static final String MOD_ID = "tau";
     public static final Logger LOGGER = LogManager.getLogger(MOD_ID);
 
-    public Tau() {
+    public Tau(IEventBus bus) {
     }
 }