+++ src/main/java/com/github/wintersteve25/tau/menu/TauContainerUI.java
@@ -1,70 +1,58 @@
-package com.github.wintersteve25.tau.renderer;
+package com.github.wintersteve25.tau.menu;
 
 import com.github.wintersteve25.tau.build.BuildContext;
 import com.github.wintersteve25.tau.build.UIBuilder;
 import com.github.wintersteve25.tau.components.base.DynamicUIComponent;
-import com.github.wintersteve25.tau.components.base.UIComponent;
 import com.github.wintersteve25.tau.layout.Layout;
-import com.github.wintersteve25.tau.theme.MinecraftTheme;
 import com.github.wintersteve25.tau.theme.Theme;
 import com.github.wintersteve25.tau.utils.SimpleVec2i;
+import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.Renderable;
 import net.minecraft.client.gui.components.events.GuiEventListener;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
+import net.minecraft.client.gui.screens.inventory.MenuAccess;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.player.Inventory;
-import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.inventory.AbstractContainerMenu;
-import net.minecraft.world.inventory.MenuType;
-import net.minecraft.world.item.ItemStack;
-import net.neoforged.neoforge.registries.NeoForgeRegistries;
-import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class MenuUIRenderer extends AbstractContainerScreen<MenuUIRenderer.Menu> {
+public class TauContainerUI extends AbstractContainerScreen<TauContainerMenu> implements MenuAccess<TauContainerMenu> {
 
-    private final UIComponent uiComponent;
+    private final TauMenu menu;
     private final List<Renderable> components;
     private final List<Renderable> tooltips;
     private final List<DynamicUIComponent> dynamicUIComponents;
-    private final List<SimpleVec2i> slots;
-    
+
     private final boolean renderBackground;
     private final Theme theme;
-    private boolean built;
 
-    public MenuUIRenderer(Menu pMenu, Inventory pPlayerInventory, UIComponent uiComponent, boolean renderBackground, Theme theme) {
+    private boolean built;
+    private int left;
+    private int top;
+
+    public TauContainerUI(TauContainerMenu pMenu, Inventory pPlayerInventory, TauMenu menu, boolean renderBackground, Theme theme) {
         super(pMenu, pPlayerInventory, Component.empty());
-        this.uiComponent = uiComponent;
+        this.menu = menu;
         this.renderBackground = renderBackground;
         this.theme = theme;
         this.components = new ArrayList<>();
         this.tooltips = new ArrayList<>();
         this.dynamicUIComponents = new ArrayList<>();
-        this.slots = new ArrayList<>();
-    }
-
-    public MenuUIRenderer(Menu pMenu, Inventory pPlayerInventory, UIComponent uiComponent, boolean renderBackground) {
-        this(pMenu, pPlayerInventory, uiComponent, renderBackground, MinecraftTheme.INSTANCE);
-    }
-
-    public MenuUIRenderer(Menu pMenu, Inventory pPlayerInventory, UIComponent uiComponent) {
-        this(pMenu, pPlayerInventory, uiComponent,true);
     }
 
     @Override
     protected void init() {
-        Layout layout = new Layout(width, height);
+        Layout layout = new Layout(menu.getSize().x, menu.getSize().y);
+        left = (width - layout.getWidth()) / 2;
+        top = (height - layout.getHeight()) / 2;
 
         components.clear();
         tooltips.clear();
         dynamicUIComponents.clear();
-        slots.clear();
-        
-        UIBuilder.build(layout, theme, uiComponent, new BuildContext(components, tooltips, dynamicUIComponents, (List<GuiEventListener>) children(), slots));
+
+        UIBuilder.build(layout, theme, menu.build(layout, theme), new BuildContext(components, tooltips, dynamicUIComponents, (List<GuiEventListener>) children(), new ArrayList<>()));
 
         built = true;
     }
@@ -75,6 +63,10 @@ public class MenuUIRenderer extends AbstractContainerScreen<MenuUIRenderer.Menu>
             this.renderBackground(graphics, pMouseX, pMouseY, pPartialTicks);
         }
 
+        PoseStack poseStack = graphics.pose();
+        poseStack.pushPose();
+        poseStack.translate(left, top, 0);
+
         for (Renderable component : components) {
             component.render(graphics, pMouseX, pMouseY, pPartialTicks);
         }
@@ -82,6 +74,8 @@ public class MenuUIRenderer extends AbstractContainerScreen<MenuUIRenderer.Menu>
         for (Renderable tooltip : tooltips) {
             tooltip.render(graphics, pMouseX, pMouseY, pPartialTicks);
         }
+
+        poseStack.popPose();
     }
 
     @Override
@@ -101,21 +95,5 @@ public class MenuUIRenderer extends AbstractContainerScreen<MenuUIRenderer.Menu>
         }
 
         super.onClose();
-    }
-
-    public static class Menu extends AbstractContainerMenu {
-        protected Menu(@Nullable MenuType<?> pMenuType, int pContainerId) {
-            super(pMenuType, pContainerId);
-        }
-
-        @Override
-        public ItemStack quickMoveStack(Player pPlayer, int pIndex) {
-            return null;
-        }
-
-        @Override
-        public boolean stillValid(Player pPlayer) {
-            return false;
-        }
     }
 }