+++ src/main/java/com/github/wintersteve25/tau/tests/ClientEvents.java
@@ -3,13 +3,13 @@ package com.github.wintersteve25.tau.tests;
 import net.minecraft.client.Minecraft;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.SubscribeEvent;
-import net.neoforged.fml.common.Mod;
+import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.neoforge.client.event.InputEvent;
 import org.lwjgl.glfw.GLFW;
 import com.github.wintersteve25.tau.Tau;
 import com.github.wintersteve25.tau.renderer.ScreenUIRenderer;
 
-@Mod.EventBusSubscriber(modid = Tau.MOD_ID, value = Dist.CLIENT, bus = Mod.EventBusSubscriber.Bus.FORGE)
+@EventBusSubscriber(modid = Tau.MOD_ID, value = Dist.CLIENT, bus = EventBusSubscriber.Bus.GAME)
 public class ClientEvents {
     @SubscribeEvent
     public static void onKeyDown(InputEvent.Key evet) {